+++ tests/testthat/test_popkin_BEDMatrix.R
@@ -13,8 +13,8 @@ if (suppressMessages(suppressWarnings(require(BEDMatrix)))) {
 
     # only repeat tests where genotypes X is input!
     
-    test_that("function returns precomputed values: get_A", {
-        expect_silent( obj <- get_A( X ) )
+    test_that("function returns precomputed values: popkin_A", {
+        expect_silent( obj <- popkin_A( X ) )
         expect_equal( obj$A, A )
         expect_equal( obj$M, M )
     })