+++ tests/testthat/test_popkin_fn.R
@@ -30,8 +30,8 @@ dimnames(Phi) <- NULL
 
 # only repeat tests where genotypes X is input!
 
-test_that("function returns precomputed values: get_A", {
-    expect_silent( obj <- get_A( Xf, n_ind = n ) )
+test_that("function returns precomputed values: popkin_A", {
+    expect_silent( obj <- popkin_A( Xf, n = n ) )
     expect_equal( obj$A, A )
     expect_equal( obj$M, M )
 })